#24 Move common functions to common module
