+++ pandas_dataclasses/pandas/__init__.py
@@ -1,0 +1,8 @@
+# submodules
+from . import common  # type: ignore
+from . import series  # type: ignore
+
+
+# aliases
+from .common import *
+from .series import *
